refactor: Move e2e tests to dedicated dir
